dataplane: move GetSupportedDataplaneFeatures into pbserverdiscovery
While integrating the new consul-server-connection-manager library into
consul-dataplane, we ran into protobuf namespace conflict issues because
both imported (and generated separate Go code for) the pbdataplane proto
package.

There are various workarounds available, but the most straightfoward is
to remove the need for both imports by moving everything the connection
manager needs into the pbserverdiscovery package, which users of the
library should never need to import directly.

This actually makes more sense anyway, because feature negotiation is
part of discovering an appropriate server.
