+++ proto-public/pbdataplane/dataplane.pb.binary.go
@@ -8,36 +8,6 @@ import (
 )
 
 // MarshalBinary implements encoding.BinaryMarshaler
-func (msg *GetSupportedDataplaneFeaturesRequest) MarshalBinary() ([]byte, error) {
-	return proto.Marshal(msg)
-}
-
-// UnmarshalBinary implements encoding.BinaryUnmarshaler
-func (msg *GetSupportedDataplaneFeaturesRequest) UnmarshalBinary(b []byte) error {
-	return proto.Unmarshal(b, msg)
-}
-
-// MarshalBinary implements encoding.BinaryMarshaler
-func (msg *DataplaneFeatureSupport) MarshalBinary() ([]byte, error) {
-	return proto.Marshal(msg)
-}
-
-// UnmarshalBinary implements encoding.BinaryUnmarshaler
-func (msg *DataplaneFeatureSupport) UnmarshalBinary(b []byte) error {
-	return proto.Unmarshal(b, msg)
-}
-
-// MarshalBinary implements encoding.BinaryMarshaler
-func (msg *GetSupportedDataplaneFeaturesResponse) MarshalBinary() ([]byte, error) {
-	return proto.Marshal(msg)
-}
-
-// UnmarshalBinary implements encoding.BinaryUnmarshaler
-func (msg *GetSupportedDataplaneFeaturesResponse) UnmarshalBinary(b []byte) error {
-	return proto.Unmarshal(b, msg)
-}
-
-// MarshalBinary implements encoding.BinaryMarshaler
 func (msg *GetEnvoyBootstrapParamsRequest) MarshalBinary() ([]byte, error) {
 	return proto.Marshal(msg)
 }
